changed declaration also

+++ omxd.c
@@ -31,7 +31,7 @@ static int vol_mB = 0;
 #define PLAYER_QUIT 1
 #define PLAYER_START 2
 #define PLAYER_TIMESTAMP 3
-static void push_event(int event, char *file, int ts);
+static void push_event(int event, char *file, long ts);
 
 int main(int argc, char *argv[])
 {
